Add files via upload
Tried many things to get the image. I've added Content-Type to the header but that doesn't do anything else. The number of bytes written to the client is the same amount as the actual image filesize (68042 bytes). There was originally an issue where not all of the buffer was being written if it was large, so that had to be fixed with a while loop.

Still getting getting net::ERR_CONTENT_LENGTH_MISMATCH